requirements.txt
runwayml>=0.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
google-genai>=1.5.0
elevenlabs>=1.0.0
Pillow>=10.0.0
pydantic-ai>=0.0.14
jinja2>=3.1.0
